requirements.txt
ipython
jupyter
pandas
numpy
tqdm
openai
ipykernel
langchain
python-dotenv
gitpython
learnosity_sdk
json5
lxml
tqdm
sympy
latex2markdown
unidecode
psycopg2-binary # NOTE: This is a precompiled version that will not always work https://www.psycopg.org/docs/install.html
matplotlib
scikit-learn
changelogs
urlextract
html2text
http3
requests_async
tiktoken
transformers
torch
pipreqs
sentencepiece
docker